Implement M2 jobqueue configuration.
